pyproject.toml
[tool.poetry]
name = "monguarte"
version = "0.1.0"
description = ""
authors = ["Cristian Camilo Perez Lopez <ccperez@soy.sena.edu.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
django-bootstrap5 = "^23.3"
django-crispy-forms = "^2.0"
crispy-bootstrap5 = "^0.7"
django-bootstrap-icons = "^0.8.3"
mysqlclient = "^2.2.0"
django-dump-load-utf8 = "^0.0.4"
django-safedelete = "^1.3.2"
django-seed = "^0.3.1"
django-select2 = "^8.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
